+++ src/ui/ozone/platform/drm/common/drm_util.cc
@@ -527,68 +527,6 @@ std::unique_ptr<display::DisplaySnapshot> CreateDisplaySnapshot(
       params.year_of_manufacture, params.maximum_cursor_size);
 }
 
-int GetFourCCFormatFromBufferFormat(gfx::BufferFormat format) {
-  switch (format) {
-    case gfx::BufferFormat::R_8:
-      return DRM_FORMAT_R8;
-    case gfx::BufferFormat::R_16:
-      return DRM_FORMAT_R16;
-    case gfx::BufferFormat::RG_88:
-      return DRM_FORMAT_GR88;
-    case gfx::BufferFormat::RGBA_8888:
-      return DRM_FORMAT_ABGR8888;
-    case gfx::BufferFormat::RGBX_8888:
-      return DRM_FORMAT_XBGR8888;
-    case gfx::BufferFormat::BGRA_8888:
-      return DRM_FORMAT_ARGB8888;
-    case gfx::BufferFormat::BGRX_8888:
-      return DRM_FORMAT_XRGB8888;
-    case gfx::BufferFormat::BGRX_1010102:
-      return DRM_FORMAT_XRGB2101010;
-    case gfx::BufferFormat::BGR_565:
-      return DRM_FORMAT_RGB565;
-    case gfx::BufferFormat::UYVY_422:
-      return DRM_FORMAT_UYVY;
-    case gfx::BufferFormat::YVU_420:
-      return DRM_FORMAT_YVU420;
-    case gfx::BufferFormat::YUV_420_BIPLANAR:
-      return DRM_FORMAT_NV12;
-    default:
-      NOTREACHED();
-      return 0;
-  }
-}
-
-gfx::BufferFormat GetBufferFormatFromFourCCFormat(int format) {
-  switch (format) {
-    case DRM_FORMAT_R8:
-      return gfx::BufferFormat::R_8;
-    case DRM_FORMAT_GR88:
-      return gfx::BufferFormat::RG_88;
-    case DRM_FORMAT_ABGR8888:
-      return gfx::BufferFormat::RGBA_8888;
-    case DRM_FORMAT_XBGR8888:
-      return gfx::BufferFormat::RGBX_8888;
-    case DRM_FORMAT_ARGB8888:
-      return gfx::BufferFormat::BGRA_8888;
-    case DRM_FORMAT_XRGB8888:
-      return gfx::BufferFormat::BGRX_8888;
-    case DRM_FORMAT_XRGB2101010:
-      return gfx::BufferFormat::BGRX_1010102;
-    case DRM_FORMAT_RGB565:
-      return gfx::BufferFormat::BGR_565;
-    case DRM_FORMAT_UYVY:
-      return gfx::BufferFormat::UYVY_422;
-    case DRM_FORMAT_NV12:
-      return gfx::BufferFormat::YUV_420_BIPLANAR;
-    case DRM_FORMAT_YVU420:
-      return gfx::BufferFormat::YVU_420;
-    default:
-      NOTREACHED();
-      return gfx::BufferFormat::BGRA_8888;
-  }
-}
-
 int GetFourCCFormatForOpaqueFramebuffer(gfx::BufferFormat format) {
   // DRM atomic interface doesn't currently support specifying an alpha
   // blending. We can simulate disabling alpha bleding creating an fb
