+++ src/ui/ozone/platform/drm/gpu/drm_overlay_validator.cc
@@ -8,6 +8,7 @@
 
 #include "base/files/platform_file.h"
 #include "ui/gfx/geometry/size_conversions.h"
+#include "ui/ozone/common/linux/drm_util_linux.h"
 #include "ui/ozone/platform/drm/common/drm_util.h"
 #include "ui/ozone/platform/drm/gpu/drm_device.h"
 #include "ui/ozone/platform/drm/gpu/drm_window.h"
